Fix sign-conversion warning in SetInternalGeomData
Cast ~0x200 to uint32_t to avoid GCC -Werror=sign-conversion.

+++ include/Geometry.hpp
@@ -660,7 +660,7 @@ public:
 		if (internal)
 			flags |= 0x200;
 		else
-			flags &= ~0x200;
+			flags &= ~uint32_t(0x200);
 	}
 
 	// SelectMesh provides a way to choose which mesh from the BSGeometryMesh list data accesessors will use.
